refactor: remove get dependency for UI
