+++ lib/app/modules/notices/presentation/pages/notice_page.dart
@@ -18,7 +18,7 @@ import 'package:ziggle/app/modules/notices/domain/entities/notice_summary_entity
 import 'package:ziggle/app/modules/notices/presentation/bloc/notices/notices_bloc.dart';
 import 'package:ziggle/app/modules/notices/presentation/bloc/reminder/reminder_bloc.dart';
 import 'package:ziggle/app/modules/notices/presentation/bloc/report/report_bloc.dart';
-import 'package:ziggle/app/modules/notices/presentation/widgets/article_body.dart';
+import 'package:ziggle/app/modules/notices/presentation/widgets/notice_body.dart';
 import 'package:ziggle/app/modules/notices/presentation/widgets/page_spinner.dart';
 import 'package:ziggle/gen/strings.g.dart';
